% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_intensity_norm.R
\name{global_intensity_norm}
\alias{global_intensity_norm}
\title{Global Intensity based Normalization}
\usage{
global_intensity_norm(log_matrix)
}
\arguments{
\item{log_matrix}{Log-Transformed Matrix}
}
\value{
Log-Transformed and Normalized Matrix
}
\description{
Global Intensity based Normalization
}
