% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_rlr_norm.R
\name{global_rlr_norm}
\alias{global_rlr_norm}
\title{Global Robust Linear Model based Normalization}
\usage{
global_rlr_norm(log_matrix)
}
\arguments{
\item{log_matrix}{Log-Transformed Matrix}
}
\value{
Log-Transformed and Normalized Matrix
}
\description{
Global Robust Linear Model based Normalization
}
