% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_norm.R
\name{mean_norm}
\alias{mean_norm}
\title{Mean Intensity based Normalization}
\usage{
mean_norm(log_matrix)
}
\arguments{
\item{log_matrix}{Log-Transformed Matrix}
}
\value{
Log-Transformed and Normalized Matrix
}
\description{
Mean Intensity based Normalization
}
