% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_norm.R
\name{median_norm}
\alias{median_norm}
\title{Median Intensity based Normalization}
\usage{
median_norm(log_matrix)
}
\arguments{
\item{log_matrix}{Log-Transformed Matrix}
}
\value{
Log-Transformed and Normalized Matrix
}
\description{
Median Intensity based Normalization
}
