% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_norm.R
\name{quant_norm}
\alias{quant_norm}
\title{Quantile based Normalization}
\usage{
quant_norm(log_matrix)
}
\arguments{
\item{log_matrix}{Log-Transformed Matrix}
}
\value{
Log-Transformed and Normalized Matrix
}
\description{
Quantile based Normalization
}
