% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smad_norm.R
\name{smad_norm}
\alias{smad_norm}
\title{Scaled Median Absolute Deviation based Normalization}
\usage{
smad_norm(log_matrix)
}
\arguments{
\item{log_matrix}{Log-Transformed Matrix}
}
\value{
Log-Transformed and Normalized Matrix
}
\description{
Scaled Median Absolute Deviation based Normalization
}
