% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsn_norm.R
\name{vsn_norm}
\alias{vsn_norm}
\title{Variance-Stabilizing Normalization}
\usage{
vsn_norm(log_matrix)
}
\arguments{
\item{log_matrix}{Log-Transformed Matrix}
}
\value{
VSN Transformed and Normalized Matrix
}
\description{
Variance-Stabilizing Normalization
}
